Align precip accumulation code with renames from main
